+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Added
+- Added `TriggerQuestionsForQuestion` query [#360]
 - Added missing tests for `PlanAuthoring` components, because the tests were not meeting coverage requirements [#339]
 - Added new `ArchivedPlanView` component that renders different versions, and simplified `DmpLandingPage` to mainly be a wrapper [#339]
 - Added new `useFormatDateWithMonth` to shared date utilities [#339]
@@ -42,6 +43,9 @@
 
 ## Updated
 - Redesigned `ProjectListItem` and project list pages (`/projects`, `/admin/projects`, `/admin/users/[userId]/projects`) with expandable plan details, skeleton loading, and GraphQL-backed plan and collaborator data
+- Updated `DisplayLogicComponent.tsx` to show condition joiner within condition groups as well as between condition groups [#360]
+- Updated `QuestionEdit` to call `TriggerQuestionsForQuestionDocument` to get the qualifying trigger questions [#360]
+- Updated `useTriggerQuestions` hook to remove currentDisplayOrder since the backend does those calculations before sending data [#360]
 - Updated plan title on landing page to expand across the full width [#339]
 - Updated `DmpLandingPage` to just use the new `PublicPlanVersionByDmpIdDocument` query, which calls on DynamoDB to get the data [#339]
 - Updated `ProjectsProjectFunding` page to display `View` button for funding if in `readOnly` mode [#246]